# AIP2 Client/src/model/Account.js
/**
 * Security Guard Accounts
 * RESTful API für AIP2
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 *
 */


import ApiClient from '../ApiClient';





/**
* The Account model module.
* @module model/Account
* @version 1.0.0
*/
export default class Account {
    /**
    * Constructs a new <code>Account</code>.
    * Ein User-Account
    * @alias module:model/Account
    * @class
    */

    constructor() {
        

        
        

        

        
    }

    /**
    * Constructs a <code>Account</code> from a plain JavaScript object, optionally creating a new instance.
    * Copies all relevant properties from <code>data</code> to <code>obj</code> if supplied or a new instance if not.
    * @param {Object} data The plain JavaScript object bearing properties of interest.
    * @param {module:model/Account} obj Optional instance to populate.
    * @return {module:model/Account} The populated <code>Account</code> instance.
    */
    static constructFromObject(data, obj) {
        if (data) {
            obj = obj || new Account();

            
            
            

            if (data.hasOwnProperty('id')) {
                obj['id'] = ApiClient.convertToType(data['id'], 'Number');
            }
            if (data.hasOwnProperty('name')) {
                obj['name'] = ApiClient.convertToType(data['name'], 'String');
            }
            if (data.hasOwnProperty('email')) {
                obj['email'] = ApiClient.convertToType(data['email'], 'String');
            }
            if (data.hasOwnProperty('password')) {
                obj['password'] = ApiClient.convertToType(data['password'], 'String');
            }
        }
        return obj;
    }

    /**
    * ID des Users
    * @member {Number} id
    */
    id = undefined;
    /**
    * Name des Users
    * @member {String} name
    */
    name = undefined;
    /**
    * Email des Users
    * @member {String} email
    */
    email = undefined;
    /**
    * Verschlüsseltes Passwort des Users
    * @member {String} password
    */
    password = undefined;








}